pyproject.toml
[project]
name = "pydanticai-bash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "vercel>=0.5.1",
    "vercel-sandbox>=0.1.0",
    # AI agent
    "pydantic-ai[openai]>=0.0.14",
    # HTTP client for Gong API
    "httpx>=0.27.0",
    # FastAPI server + SSE
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
